Remove incorrect detection of installed patches.
The package_patch_installed_regex is used to identify patches that
have already been installed, but that is not what yum does when run
with "check-update", it list patches (which are just packages) which
have *not* been installed. Therefore remove this identification and
assume that all listed packages are available patches that can be
installed.