Don't use package promises in test initialization phase.
It's better to use explicit commands so that we don't risk the test
passing because the initialization did not work correctly.

This revealed a few new failures, some of which were fixed by tweaking
the package promise constraints, but one test case could not be fixed
without breaking the test case premise, so it was moved to staging
(associated with Redmine #1535).