RM4890: Put quotes round non-function edit_line names.
It's not clear that we *should* require these, but I get errors, in
their absence, about "Apparent body '$name' was undeclared, but used
in a promise near $loc (possible unquoted literal value)" and these
quotes fix the error.  See RedMine 4890 for discussion of whether we
should cope without the quotes.