Skip package tests on Redhat 4.
RPM on Redhat 4 has a bug which corrupts the RPM database when we do
our tests. The cause is unknown, but it is suspected to have something
to do with the fact that we switch out the official repository with a
custom one. Either way, it renders package promises untestable on that
platform.