Escape the dots in def.domain where used in def.acl
I saw
 admit_hostnames: .*intra.cfengine.com
in my hub's report of what it would grant access to.  The dot at the
start is clearly intended as a regex so I must suppose the later dots
are treated as such, too.  So the owner of YYYintraXcfengine.com can
get at my hub; this probably isn't prudent.  So escape the dots in the
domain and put an (escaped) dot before it, so that only hosts within
the specified domain are granted access.