Change: Make paths bundle dynamic
DO NOT MERGE: THIS IS ONLY FOR CONSIDERATION AND FEEDBACK

Instead of having a list of each path that we want to know about we
could find bins based on a pre defined path, just like PATH behaves on
the cli.

This is risky, possibly too risky. I am unsure if this is offset by
making the paths bundle more usable.

Additionally some bins have characters in the name that make them
generate erros. This could probably be worked around.

Currently this implementation relies on GNU find. That may be an
unacceptable dependancy.