Change: Enterprise HA policy cleanup
Ref: https://dev.cfengine.com/issues/6813

The CFEngine Enterpise HA policy uses features from 3.6. When the policy
is loaded on 3.5 and earlier clients it fails to validate.

Make the HA policy only load on En Additionally there is no reason to
load the HA policy on clients that are not enterprise hubs.

Change class names to make them more self describing and specific to
reduce changes of clashing with custom policies.

Move enablement of hub ha to def and update_def from ha_def with other
feature enablement classes.