Change: Reorganize cfe_internal
Not all policy that ships with CFEngine is applicable or even useful to
community users. This work is to seperate the enterprise specific policy
from the core policy and only load the enterprise specific policy when
necessary.

While these changes are compatible with 3.6 but will likely be kept in
master until 3.7 due to it being a large change.