Change: Remove redundant package promise
There is no need to have two package promises to upgrade Enterprise. I
chose to remove the package promises guarded by stopped_cfprocs because
in order for those promises to be activated the software must be
downloaded from the hub, and cfengine processes must be stopped within
the same agent run. This creates a brittle chain.

Ref: https://dev.cfengine.com/issues/6964