More FreeBSD fixes...
Meet the Major Version Concatenator; a little pain here is a lot easier than having to add some FreeBSD specific code to core - for a hardclass that would only affect one operating system. Note that the FreeBSD MVC is basically *REQUIRED* for packages.cf fixup for pkgng and this seemed like the best place to put it, since it also cleans up paths.cf. I'm certainly open to better suggestions as to method and/or placement - this just seemed the least intrusive solution.

Other notable fixes...
* Attempt to address perl <5.20 being at /usr/bin/perl and >=5.20 NOT having /usr/bin/perl with an slist. Tested okay here, but more eyes are better.
* Add setfacl (POSIX.1e), getfmac (Mandatory Access Control), setfmac - Redmine #6972 appears to be MAC related.
* Take another swing at path[ping] since transition is 9.2-pSomething->9.2-pSomethingHigher / 9.3
* Added `path[pkg]` for FreeBSD 9/10/11 _and_ DragonflyBSD which also uses it.