Fix package version parsing for pkgin-0.8.0.
Pkgin version 0.8.0 and later, when STDOUT is not a terminal separates
the package name/version from the description with a semi-colon. Prior
to 0.8.0 the description was separated by whitespace.

E.g.:

New behavior:

    zsh-5.0.7nb1;The Z shell

Old behavior:

    zsh-5.0.5nb1         The Z shell

This would cause cfengine to interpret the version as `5.0.7nb1;The`.

This backward compatible fix adjusts the package_version_regex for the
`smartos` package method to correctly match the version number.