Remove: Extra def.json and yaml from 3.6 upate inputs
Extra def.json in lib only adds extra complexity. It may make sense in
the future to merge def.json with something outside of masterfiles (like
data sourced from a CMDB) but for now it seems to add little.

3.6 has no yaml support, so probably it makes little sense for 3.6
agents to download YAML. Besides it can be overridden in def.json now so
even if a user wants it, it wont require a change in policy.