Change: Improve CPU info inventory
dmidecode is unreliable, sometimes it outputs very badly formated
processer version information, sometimes it cannot get processor
information at all. /proc/cpuinfo is a better source (at least on
linux), and windows doesnt use dmidecode, it uses another method but
offers the data as if it were from dmidecode.

This change abstracts cpuinfo into its own bundle. If /proc/cpuinfo is
avaialable it uses that (based on parsing proc), and if only dmidecode
is available it will use that.

This does not completly detangle windows from dmidecode, it only
detangles CPU model name from being inventoried directly in the
dmidecode bundle. This can serve as initial work to continue detangling
other information gathered in the dmidecode bundle that is not actually
from dmidecode.

Additionally there is stubbed out support for inventorying
hyperthreading. It is left commented out as its detection is not
completely straight forward, and it has questionable value.

Ref: https://dev.cfengine.com/issues/7014