Change: Re-organize controls
Complete split of controls and relocate update_def.cf into controls/VER
for consistency.

Switch to classic array definition of control inputs to make it easier
to exclude extra input files that are not necessary on community
installs and make the function of each control input more clear within
the policy.