Re-fix broken upgrade procedure failing on Windows.
This had been fixed already in cff8ff50e6, but that only partially
fixed it, since there was a different section with the same bug in it
elsewhere. The bug was reexposed in 403af5633e, when the fixed part
was cleaned up and removed and the buggy part remained.
(cherry picked from commit 284dab9d6baaf48ef500539981ece73f35fb26c4)