Add apt-get and yum package module tests.
Currently only apt is covered in the tests.

The tests are organized in numbered directories simply to make the
timed tests which are not unsafe run in parallel with other tests
which are not unsafe.

Redmine #7121 and #7122.
(cherry picked from commit f34ebd0b77d1eaf3421483bc6a4518d776d2ee63)