Fix repository paths becoming too long in Jenkins.
The paths become gigantic when combined with all the labels and
workspaces in Jenkins, so copy the repository to a shorter location
before using it. These are unsafe tests so we are free to put it
wherever we want.
(cherry picked from commit 67497e20af27135d51798447b7b9355cd687873a)