Fix: Superflous reporting about chkconfig usage
Inform mode is supposed to be for changes, previously a report was
printed each time the agent was run in inform mode even if there were no
changes. This guards the chkconfig report so that it is only printed
under the same conditions that trigger the command to be run.

Ref: https://dev.cfengine.com/issues/7196
(cherry picked from commit 0a7464de936e101a45a0ea6fe373dfbaec9cca9d)

Conflicts:
	lib/3.7/services.cf