Fix: standard_services restart when service not running with chkconfig
Previously if a service was not running and policy restart was specified
the service would not be restarted (and an inform level report would
indicate that it was). Now the service is restarted even if its not
running.

Ref: https://dev.cfengine.com/issues/7258
(cherry picked from commit 1f48a2ca6763fccdd898b9708c6990e61ab3641d)

Conflicts:
	lib/3.8/services.cf