Correctly detect if systemd is used in Debian
CFEngine 3.6 tries to understand if a Linux is using systemd as init system
by looking at the contents of /proc/1/cmdline, that happens in
bundle common inventory_linux. That's indeed a smart thing to do but unfortunately
fails on Debian Jessie, where you have:

root@cf-test-v10:~# ls -l /sbin/init
lrwxrwxrwx 1 root root 20 May 26 06:07 /sbin/init -> /lib/systemd/systemd

the pseudo-file in /proc will still report /sbin/init and as a result the systemd
class won't be set. This affects services promises negatively

This patch checks if /sbin/init is a link; if it is, and it points to systemd,
the systemd class is set