Add: Comment about RHEL4 dmidecode based inventory
We learned that there are package updates available that allow RHEL4 to
be used with the dmidecode inventory for redhat 4. Automatically
detecting the package condition is considered too complex compared to
the value provided so it is left for the user to adjust as necessary.

(cherry picked from commit bf02336dc830eb219dacd726a7f9a288d0c64b4f)