Change: Switch inform_mode reports to DEBUG|DEBUG_bundlename
Inform mode is intended to show changes to the system. The current use of
inform_mode throughout the framework can lead to much more information than
needed.

Using a DEBUG|DEBUG_bundlename class guard allows this policy level debug
information to be reported for the entire policy run or for individual bundles
in a consistent way.

Ref: https://dev.cfengine.com/issues/7197