Change: Switch inform_mode reports to DEBUG|DEBUG_bundlename
Inform mode is intended to show changes to the system. The current use of inform_mode throughout the framework can lead to much more information
than needed.

Using a DEBUG|DEBUG_bundlename class guard allows this policy level debug information to be reported for the entire policy run or for individual
bundles in a consistent way.

Ref: https://dev.cfengine.com/issues/7197 (cherry picked from commit 6f58db5b10a3fc84b8519133ab669e48293f13b7)

Conflicts:
	CHANGELOG.md
	inventory/any.cf
	lib/3.8/commands.cf
	lib/3.8/feature.cf
	lib/3.8/files.cf
	lib/3.8/packages.cf
	lib/3.8/services.cf
	lib/3.8/vcs.cf