Revert addition of services_autorun class guard in promises.cf.
Although it looks like it should work, the class is somehow defined
too late for it to work, and the promises/autorun.cf test fails with
this. This may need additional restructuring to make sure the class is
defined at the right moment.

The services_autorun class guard is a recent addition, so this should
not break any existing version.