Change: Improve git drop user support
When using bundle agent git from the stdlib on repositories that are not
owned by root an error prevents action because cfengine does not inherit
the shell environment of the user it is executing as. We now set the
HOME variable when running git commands with dropped privledges.

Changelog: Title