CFE-2183: Figure out version instead of hardcoding.
This way we can build all versions we want automatically, without
having to hardcode the version number in the source, using the
EXPLICIT_VERSION variable. If the variable is not set, and git is
available, it will try to figure it out from git tags.

Note that all of this is done in the autogen stage, because git will
generally not be available at the configure stage. This is also
required in order for tarballs to come with the correct version
embedded in them, because they won't have git available to figure this
out.

Note that some files have been given the .in suffix in order to have
version strings replaced in them, using standard AC_CONFIG_FILES
autoconf style. There are a lot more version strings around, but they
are not important for the functioning of CFEngine and can be
considered later.
(cherry picked from commit bcfc5c169a8ea3071b42fee517f9e65079e32adb)

Conflicts:

	.travis.yml
	configure.ac
	controls/3.6/update_def.cf.in
	controls/update_def.cf.in
	promises.cf.in