Change: Autorun Tests
It is expected that if the class services_autorun is defined from policy
that it will not be defined soon enough to automatically extend inputs
with the policy files found in services/autorun.

Additionally add a test that when services_autorun is defined from the
augments file it works as expected similar to how we have a test for
defining services_autorun as an option to the agent.