Make Database purge schedule more robust
Rather than restricting the database maintenance to specifically
the first 5 minutes of the midnight - 1AM hour, restrict it only
to some point within that hour, and use 'if_elapsed_day' to
prevent multiple runs in a single day.

With the default execution schedule and default splaytime, this
will be effectively the same as before, but this change will
make the DB maintenance promises more robust against execution
schedule changes.  (E.g. if CFEngine only runs every 20 minutes
with an splaytime to match, it may never run in the Min00_05
interval and thus never perform the DB maintenance needed.
This commit fixes that.)