Change: Refactor runalerts for systemd
Systemd has issues managing a bash script as a service always seeing
that it had died even when it was running. The script was re-implemented
in php which bypasses the issues with systemd and the way processes are
detected.

(cherry picked from commit 2ff0609886abc71fe5f481f35412d667dc3e5fe3)