Add enable/disable repo options to yum package module
This is implemented for the 'list-updates' command,
the 'list-updates-local' command, and the 'repo-install'
command (from the package module API).

(https://docs.cfengine.com/docs/3.10/reference-package_modules.html)

The option is not applicable to the other API commands
(supports-api-version, get-package-data, list-installed,
file-install, remove).

Illustration of usage for package installation:

  packages:
      "some-odd-package"
        policy => "present",
        options => { "enablerepo=CUSTOM" };

Usage for inventory:

In order to use the enablerepo or disable repo options
for package inventory, you need to make a copy of the
package module and define an additional package_module body
in your policy (see CFE-2103 for why this kludge is
necessary).  (That is, unless you wish for the options
to be used for every package promise you make, as well
as being used for inventory.)

Example inventory usage:

Call the copy of the package module 'yum_inventory'

Define the package module body like so:

body package_module yum_inventory {
  default_options => { "disablerepo=*", "enablerepo=UPDATES" };
}

Define the yum_inventory package module as the one to use for inventory:

body common control {
  package_inventory => { "yum_inventory" };
}

Your "available updates" inventory will now be done using your
"UPDATES" yum repository only, even if that repo is usually disabled.
All your package promises will be unaffected.

Note: in the original implementation I passed a tuple
to the startswith() method:

    if option.startswith(("enablerepo=","disablerepo=")):

However, this throws an error on the Python version that
ships with RHEL 5, so I switched to using 'or' instead.

Thanks to Herman Schumacher for recommending line.strip()
and the more readable 'option = line[len("options="):]'
(to replace the obscure 'option = line[8:]').

(cherry picked from commit 719e9c658596d644643722d32c5d30a33d3a3ebb)