Fix rare bug that would sometimes prevent redis-server from launching.
redis-server will not launch if run from a directory it cannot read.
This is almost never the case, since CFEngine generally chdirs to "/"
as part of other promises. This is not guaranteed, however, and in
certain rare circumstances it may not happen. This was observed if
certain promises were removed.

Probably this bug will never happen in production, but at least this
way we are 100% sure.

Changelog: Title
Signed-off-by: Kristian Amlie <kristian.amlie@cfengine.com>
(cherry picked from commit da619d281c682c8bfefc119bbebca7fb7a42dd1f)