ENT-3328: Ensure HA standby hubs have am_policy_hub state marker
The am_policy_hub state marker is created when a host is bootstrapped to an ip
found bound to one of its interfaces. When standby hubs are bootstrapped to the
primary this state file is not created automatically and standby hubs may be
unable to determine their hub status. This change ensures that any host that
finds the IP of an HA hub member will have an am_policy_hub marker found in
state.

Changelog: Title