ENT-3390: Fix to frequent logging of enterprise license utilization
This change prevents logging of Enterprise license utilization more frequently
than desired. It was being logged on each agent run instead of waiting for the
expected period of time to pass before logging triggered again. The promiser
contains $(sys.date) which is different for each agent run making the promise
unique on each run and the lock useless. Locking was traded out for a persistent
class which is defined and communicates that logging has happened. When the
class expires the promise will actuate again and the resulting class will
persist guarding it until the class expires.

Changelog: Title