Sort indices in insert_ini_section()
The order of the key-value pairs doesn't matter in INI files, but
they are easier to read if it they are sorted by key.

Also make sure indices are sorted in the insert_ini_section()
test. Otherwise they may be in arbitrary order.

Use reverse(sort()) to make sure the sections in the resulting
file are alphabetically ordered (in combination with how
insert_ini_section() works).