ENT-3603 Fix self upgrade for rpm packages with default names
In 3.6.6 CFEngine published packages began containing a platform
identifier (el4, el6) to identify which platform a package was optimized for.
This allowed for users to download packages into the same directory without
overwriting the same package for an older platform.

This change aligns the filename the self upgrade policy looks for to include
this platform identifier.

Changelog: Title