ENT-3659 Localize delete tidy in ha update policy
This fixes an issue where on an ha enabled hub the update policy errors because
it was using a definition of the bundle from stdlib in the update policy which
does not include the stdlib. To correct the error I copied the body locally and
renamed it to ha_tidy.

Changelog: Title
(cherry picked from commit 77729a78742642193c45f996e148a9689fffe010)