Change class identifying runagent initiated executions from cfruncommand to cf_runagent_initated
Changelog: Title

After further testing and discussions with Jimis we determined that in 3.7.x if
`cfruncommand` is found in the string of an agent execution, it disallows
specification of inputs.

The error was coming from the first half of the `cfruncommand` related to policy
update, not from the automatically appended `-I -Dcfruncommand`.

History:

`cfruncommand` when introduced was intended to be a single command. You can
still find traces of this in the early commits of masterfiles where it had a
default value of just `$(sys.cf_agent)` and was commented out.

In 2012 Jeang seems to have introduced the addition of the update policy in the
default `cfruncommand` with commit a67db2005e5bcd2d4544ba5bc7e683a67bd0b809.

In 3.7.x `-I -Dcfruncommand` was appended to the end of `cfruncommand`.

The class `cfruncommand` was intended to communicate to the policy initiated by
`cf-runagent` that it was initiated out of turn. This allows for the policy to
determine if it should do something based on the initialization
source (scheduled from `cf-execd`, out of turn from `cf-runagent`).

So to work around the history related to `cfruncommand` disallowing additional
inputs and to preserve the ability for runagent initated policy to know that it
was indeed indeed from `cf-runagent` I have changed the definition of the
`cfruncommand` class to `cf_runagent_initated`.

(cherry picked from commit 18abe7b260d9d226bf9178fff157fba56919736a)