ENT-3910 Fix policy backward compatibility
In newer versions of cfengine the promise to run the bundles would have been
skipped if there were no bundles to run. This makes that behavior explicit, and
compatible with 3.7.0 to avoid an error about trying to actuate a bundle that
does not exist.

Additionally the sort function expects the name of a list, not a list itself to
be passed as the first parameter. In later versions the agent accepts the list
notation as well.