ENT-4807 Cleanup old report data relative to the most recent changetimestamp
Instead of purging all records older than the last report timestamp we now purge
data relative to the most recent changetimestamp for the host in that table.

Since the report collection allows for each report type to fail separately
this ensures that we keep the most recent data even if it is very old but the
host has newer data in a different table.

Thanks for the suggestion Mike Weilgart