CFE-2932 Detect systemd service enablement for non native services
Because non-native services have UnitFileState=bad we need an alternate way to
detect if the service is enabled or not. Since systemctl is-enabled
logs the fact that the service is not native to stderr, the probe must
be run in a shell and the output must be redirected to /dev/null or the
agent will output the information.

Changelog: Title