Skip the license utilization logging bundle according to persistent class
If the persistent class says we logged it, don't call the bundle as
hostsseen() and hostswithclass() are both potentially expensive function
calls.  Without this commit, both function calls are performed on every
agent run on hubs with license utilization logging enabled.  (The
reports promise within the bundle is guarded against evaluation every
run and only runs if the persistent class isn't set, so that the log
file doesn't grow without bound, but the *expensive* part is running the
function calls.)