Trust the methods promise that calls license logging
Nowhere else in the MPF is there a WARNING promise like the one I
removed here, that tells you a bundle was actuated but not enabled.
It's redundant, and it makes the class expressions much harder to read.
The bundle is only called in case the class is defined, and that should
be sufficient.

(Leaving in the sanity checks to only run this on an enterprise policy
hub is different, since that's the only place this bundle would make
sense to run, ever.)