Reorganize the contexts in log_cfengine_enterprise_license_utilization
Avoid running the hostsseen() or hostswithclass() functions unless the
results are needed either for (a) reporting into the log file, or (b)
reporting as part of DEBUG reporting.

I don't know that it's really necessary to have this as part of DEBUG
reporting; it would be simpler to just remove that DEBUG reporting
promise so that the logic flow would be simpler: just don't do anything
with the bundle so long as the persistent class is defined.