Turn off debug reporting of the license usage logging unless asked for
Even running this in inform mode or general DEBUG mode is a bit too
much, as by evaluating the hostsseen and hostswithclass functions the
performance is *markedly* affected on heavily populated hubs.  So if
you're trying to debug something *other* than this bundle, you would get
results obscured by the performance difference caused by evaluating
these functions.  Best not to do that unless this bundle is what you're
actually debugging.