CFE-2932 Update policy can now skip local copy optimization on policy servers
Changelog: Title

Policy servers avoid making a local connection to cf-serverd during policy
updates. cf-serverd is not always running, so this allows policy updates for the
hub itself, even when cf-serverd is not available (like during bootstrap). For
testing it is desirable to be able to excercise network access over the local
network. With this change if the mpf_skip_local_copy_optimization class is
defined, then policy update will happen over the local network.