Fix automatic version detection after tagging 3.13.0
Our build system version detection is based on assumption that if we
are on any 3.[0-9].x branch, then we're not on master branch. When
3.13.0 branch points to a commit in master branch and doesn't have
any extra commits, it wrongfully assumes that we're on this branch
even if we're actually on master. This empty commit fixes it by moving
3.13.0 branch one step away from master.

Changelog: None

Signed-off-by: Aleksei Shpakovskii <aleksei.shpakovskii@cfengine.com>