ENT-4362 Purge collection status records from the future
Changelog: Title

An issue with clocks, or client side database corruption can cause the
timestamps in the status table to become huge. This creates a problem patching
reports as the timestamps no longer reflect real ranges. This change ensures
that any timestamps for report collection status records that are in the future
will be purged, thus maintaining the ability to properly calculate deltas
between time ranges.