Avoid calling yum at every agent run
As it was part of a common bundle, it was executed everytime even when
not using old package promises. Furthermore, this simple yum command lead
to contact the server in some cases. As the option was introduced in RHEL 6.1,
use the RHEL version instead of dynamic detection.

(cherry picked from commit b05fb41a6dcaa7ac07e54aef12fd884c3c5c9ea7)