Add support for pull to FR/transport.sh
Making the superhub pull the dump files from the feeder hubs is a
better strategy than pushing the files from the feeder hubs to
the superhub. The superhub knows better when it has a capacity to
process the files and also, there's only one SSH private key that
has to be kept safe -- the superhub's key.

However, in some situations the push approach may be necessary so
let's not just delete the functionality. The transport.sh script
can easily support both modes.

Ticket: ENT-4499
Changelog: None