Guard loading cmdb data on presence
Guarding on "OK" would guard to file repaired, or exists as same upstream, it
would fail to load if there were some error in updating the file. Instead, load
it if it's there. Also, no need to define classes for the files promise, no
classes are being leveraged, nor are they tagged for reporting to enterprise
hub. Plus, it's OK if there is not a host specific data file, not all hosts have
host specific data.