Use explicit knowledge to determine if role is superhub
This change switches identification of superhub role to be based on explicit
setting in the configuration file (federation-config.json). Prior to this
change, knowledge of superhub role was based implicitly on the fact that there
were at least one configured feeder hub enabled.

This change was necessary because you can not configure a feeder hub until after
the superhub has been enabled and configured.