Refresh inventory after import on superhub
Refreshing inventory on a superhub with tens of thousands of
hosts can easily take a couple minutes and thus should not be
part of every agent run. It's also not necessary to refresh the
inventory on a superhub too often because reporting data only
changes when successfully imported from the feeder hubs. The only
exception is the superhub itself which is reporting to itself
with the standard reporting interval.

Let's just fully refresh the inventory after a successfull data
import on the superhub and only refresh the inventory data for
the superhub itself periodically (as part of agent run).

Ticket: ENT-4527
Changelog: None