Fixed inventory for Memory size (MB)
isvariable( $(variable) ) is not the same as isvariable( "variable" )

In the first case, we are checking if the resolved value of variable is a defined
variable, in the second case we are checking if variable is a defined variable.

This is a fairly common mistake. Not sure how we can improve the situation for
policy writers because both are completely valid and useful in different
contexts.

Ticket: ENT-4637
Changelog: None