Set default access promises for directories to only share if directory exists
In particular, /var/cfengine/templates does not exist by default. Running
cf-serverd with log level info emits messages noting the path does not exist.

info: Failed to canonicalise filename '/var/cfengine/templates/' (realpath: No such file or directory)
info: Path does not exist, it's added as-is in access rules: /var/cfengine/templates/

Instead of sending a user chasing to see if something is wrong, we now just
avoid providing access to paths that are not directories when we expect them to
be.

Ticket: CFE-3060
Changelog: Title