Made FR policy set up transport user regardless of role
This is necessary since there is no enable feeder API.
Feeders become enabled, but lack a role in the config.

This fixes an issue where parts of the policy assumes that the
cfdrop user has been created, while it will in some cases not
exist, because am_feeder and am_superhub are both undefined.

Ticket: ENT-4674
Changelog: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@cfengine.com>